# src/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Express Training API",
      "description": "Express Training API with Swagger",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:9000",
    "basePath": "/",
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
      "/addUser": {
        "post": {
          "tags": [
            "Trainee"
          ],
          "description": "Create new user in system",
          "parameters": [
            {
              "name": "user",
              "in": "body",
              "description": "User that we want to create",
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "New user is created",
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          }
        } 
      },
      "/users": {
        "get": {
            "tags": [
              "Trainee"
            ],
            "summary": "Get all users in system",
            "responses": {
              "200": {
                "description": "OK",
                "schema": {
                  "$ref": "#/definitions/Users"
                }
              }
            }
          }
      },
      "/user/{id}": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of user that we want to find",
            "type": "integer"
          }
        ],
        "delete": {
          "summary": "Delete user with given ID",
          "tags": [
            "Trainee"
          ],
          "responses": {
            "200": {
              "description": "User is deleted",
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          }
        },
        "put": {
          "summary": "Update user with give ID",
          "tags": [
            "Trainee"
          ],
          "parameters": [
            {
              "name": "user",
              "in": "body",
              "description": "User with new values of properties",
              "schema": {
                "$ref": "#/definitions/updateUser"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "User is updated",
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      }
    },
    
    "definitions": {
      "User": {
        "required": [
          "name",
          "_id",
          "companies"
        ],
        "properties": {
          "_id": {
            "type": "integer",
            "uniqueItems": true
          },
          "isPublic": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "books": {
            "type": "array",
            "items": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  }
                }
            }
                
          },
          "companies": {
            "type": "array",
            "items": {
                "type": "string"      
            }
                
          }
         
        }
      },
      "updateUser": {
        "required": [
          "name",
          "companies"
        ],
        "properties": {
          "isPublic": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "books": {
            "type": "array",
            "items": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  }
                }
            }
                
          },
          "companies": {
            "type": "array",
            "items": {
                "type": "string"      
            }
                
          }
         
        }
      },
      "Users": {
        "type": "array",
        "$ref": "#/definitions/User"
      }
     
    }
  }